refactor(desk-tool): add package exports directory and add proper type exports
