+++ packages/@sanity/desk-tool/src/_exports/index.ts
@@ -1,0 +1,2 @@
+export * from '../constants'
+export * from '../contexts/PaneRouterContext'
