+++ packages/@sanity/desk-tool/src/_exports/structure-builder.ts
@@ -1,0 +1,1 @@
+export {default} from '../structure-builder'
