--- packages/@sanity/desk-tool/src/index.js
@@ -1,2 +1,0 @@
-export * from './constants'
-export * from './contexts/PaneRouterContext'
