+++ packages/@sanity/desk-tool/src/structure-builder.ts
@@ -1,0 +1,1 @@
+export {StructureBuilder as default} from '@sanity/structure'
