+++ packages/@sanity/desk-tool/structure-builder.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-module.exports = require('@sanity/structure').StructureBuilder
+require('./_interopReexport')(module.exports, require('./lib/_exports/structure-builder'))
